updated flag placement
