+++ cmd/gossamer/main.go
@@ -18,10 +18,11 @@ package main
 
 import (
 	"fmt"
+	"os"
+
 	"github.com/ChainSafe/gossamer/cmd/utils"
 	log "github.com/inconshreveable/log15"
 	"github.com/urfave/cli"
-	"os"
 )
 
 var (
